requirements.txt
flask>=1.0
flask-sqlalchemy>=2.3.2
flask-wtf>=0.14.2
cymysql>=0.9.1
flask-cors>=2.1.0
flask-httpauth>=2.7.0
requests>=2.18.4
psycopg2>=2.8.3
wtforms>=2.2.1